fix the potential overflow related to B*T*V

+++ dev/cuda/common.h
@@ -35,7 +35,7 @@ void cublasCheck(cublasStatus_t status, const char *file, int line)
 // ----------------------------------------------------------------------------
 // random utils
 
-float* make_random_float_01(int N) {
+float* make_random_float_01(size_t N) {
     float* arr = (float*)malloc(N * sizeof(float));
     for (int i = 0; i < N; i++) {
         arr[i] = ((float)rand() / RAND_MAX); // range 0..1
@@ -43,7 +43,7 @@ float* make_random_float_01(int N) {
     return arr;
 }
 
-float* make_random_float(int N) {
+float* make_random_float(size_t N) {
     float* arr = (float*)malloc(N * sizeof(float));
     for (int i = 0; i < N; i++) {
         arr[i] = ((float)rand() / RAND_MAX) * 2.0 - 1.0; // range -1..1
@@ -51,7 +51,7 @@ float* make_random_float(int N) {
     return arr;
 }
 
-int* make_random_int(int N, int V) {
+int* make_random_int(size_t N, int V) {
     int* arr = (int*)malloc(N * sizeof(int));
     for (int i = 0; i < N; i++) {
         arr[i] = rand() % V; // range 0..V-1
@@ -59,13 +59,13 @@ int* make_random_int(int N, int V) {
     return arr;
 }
 
-float* make_zeros_float(int N) {
+float* make_zeros_float(size_t N) {
     float* arr = (float*)malloc(N * sizeof(float));
     memset(arr, 0, N * sizeof(float)); // all zero
     return arr;
 }
 
-float* make_ones_float(int N) {
+float* make_ones_float(size_t N) {
     float* arr = (float*)malloc(N * sizeof(float));
     for (int i = 0; i < N; i++) {
         arr[i] = 1.0f;
